refactor(tau-tools): extract runtime helpers from oversized tools.rs (#1633)
